// apps/extension/src/pages/nfts/generalized/types.ts
export type NFTInfo = {
  token_uri: string
  extension: unknown
}

export type NFTDisplayInformation = {
  name: string
  image: string
}

export type CollectionInfo = {
  name: string
  symbol: string
}

export type NFTDetailedInformation = {
  imgSrc: string
  name: string
  collection: string
  description: string
  tokenUri: string
  collectionAddress: string
  tokenId: string
}